% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge2camelot.R
\name{bridge2camelot}
\alias{bridge2camelot}
\title{Convert bridge lables to camelot metadata}
\usage{
bridge2camelot(
  photo_path,
  report_path,
  report_name = paste0("bridge2camelot-", Sys.Date()),
  maxentries = 65535,
  extensions = c("JPG", "JPEG", "PNG"),
  progress = TRUE,
  verbos = TRUE
)
}
\arguments{
\item{photo_path}{where to find the files}

\item{report_path}{path to save the camelot metadata}

\item{report_name}{a string, the base file name of the reports, default the date}

\item{maxentries}{maximum entries in one csv file, default 65535}

\item{extensions}{photo extensions, default c("JPG","JPEG","PNG"), case insensitive}

\item{progress}{whether to have a progress bar, default true}

\item{verbos}{whether to report during the progress}
}
\value{
a bool return on whether there are errors, also with warning and error log if there is any
}
\description{
This function will read photos metadata, convert it into a camelot report as good as we could
}
\details{
Please check the package exifr for instructions on how to install needed dependencies, especially see ?exifr::read_exif
}
