% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camelot2XMP.R
\name{camelot2xmp}
\alias{camelot2xmp}
\title{write camelot entries to XMP via exifr}
\usage{
camelot2xmp(
  camelot,
  errorlog_path = getwd(),
  delete_original = FALSE,
  progress = TRUE,
  verbos = TRUE
)
}
\arguments{
\item{camelot}{data.frame, the camelot data with at least Absolute.Path column}

\item{errorlog_path}{string, the direction to store the error log (if there is any), default is the working directory}

\item{delete_original}{boolean, the exiftool will automatically creat a copy of the original photo end with "_original", do you want to delete it? The default is not to delete it (FALSE)}

\item{progress}{boolean, whether to have a progress bar}

\item{verbos}{boolean, whether to be verbos}
}
\value{
boolean, whether we encounter errors
}
\description{
This function read a camelot data.frame and write the metadata into XMP part of the Jpeg file as a JSON string
}
\details{
Make sure that exiftool are proper configed, check documentation in exifr and exiftool https://exiftool.org/. Especially when on Windows, you might need to install Perl. The tag written is xmp-dc:Subject, see exiftool's documentation for detail of the tag https://exiftool.org/TagNames/XMP.html
}
